+++ cashservice/src/main/java/com/cashservice/dto/CashRequestDto.java
@@ -5,10 +5,12 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 
+import java.math.BigDecimal;
+
 public record CashRequestDto(
     @Email String email,
     @NotNull Long accountId,
     @NotBlank String currency,
-    @Positive Double amount,
+    @Positive BigDecimal amount,
     @NotNull boolean isDeposit
 ) {}
