+++ exchangeservice/src/main/java/com/exchangeservice/dto/ConversionRateDto.java
@@ -1,5 +1,7 @@
 package com.exchangeservice.dto;
 
+import java.math.BigDecimal;
+
 public record ConversionRateDto(
-    Double rate
+    BigDecimal rate
 ) {}
