+++ exchangeservice/src/main/java/com/exchangeservice/dto/ExchangeRate.java
@@ -1,7 +1,9 @@
 package com.exchangeservice.dto;
 
+import java.math.BigDecimal;
+
 public record ExchangeRate(
     String title,
     String currency,
-    Double value
+    BigDecimal value
 ) {}
