+++ transferservice/src/main/java/com/transferservice/dto/TransferRequestDto.java
@@ -5,11 +5,13 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 
+import java.math.BigDecimal;
+
 public record TransferRequestDto(
     @Email String email,
     @NotNull Long senderAccountId,
     @NotBlank String senderAccountCurrency,
-    @Positive Double amount,
+    @Positive BigDecimal amount,
     @NotNull Long recipientAccountId,
     @NotBlank String recipientAccountCurrency
 ) {}
